Struct aws_sdk_firehose::client::fluent_builders::CreateDeliveryStream

pub struct CreateDeliveryStream<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Fluent builder constructing a request to CreateDeliveryStream.

Creates a Kinesis Data Firehose delivery stream.

By default, you can create up to 50 delivery streams per AWS Region.

This is an asynchronous operation that immediately returns. The initial status of the delivery stream is CREATING. After the delivery stream is created, its status is ACTIVE and it now accepts data. If the delivery stream creation fails, the status transitions to CREATING_FAILED. Attempts to send data to a delivery stream that is not in the ACTIVE state cause an exception. To check the state of a delivery stream, use DescribeDeliveryStream.

If the status of a delivery stream is CREATING_FAILED, this status doesn't change, and you can't invoke CreateDeliveryStream again on it. However, you can invoke the DeleteDeliveryStream operation to delete it.

A Kinesis Data Firehose delivery stream can be configured to receive records directly from providers using PutRecord or PutRecordBatch, or it can be configured to use an existing Kinesis stream as its source. To specify a Kinesis data stream as input, set the DeliveryStreamType parameter to KinesisStreamAsSource, and provide the Kinesis stream Amazon Resource Name (ARN) and role ARN in the KinesisStreamSourceConfiguration parameter.

To create a delivery stream with server-side encryption (SSE) enabled, include DeliveryStreamEncryptionConfigurationInput in your request. This is optional. You can also invoke StartDeliveryStreamEncryption to turn on SSE for an existing delivery stream that doesn't have SSE enabled.

A delivery stream is configured with a single destination: Amazon S3, Amazon ES, Amazon Redshift, or Splunk. You must specify only one of the following destination configuration parameters: ExtendedS3DestinationConfiguration, S3DestinationConfiguration, ElasticsearchDestinationConfiguration, RedshiftDestinationConfiguration, or SplunkDestinationConfiguration.

When you specify S3DestinationConfiguration, you can also provide the following optional values: BufferingHints, EncryptionConfiguration, and CompressionFormat. By default, if no BufferingHints value is provided, Kinesis Data Firehose buffers data up to 5 MB or for 5 minutes, whichever condition is satisfied first. BufferingHints is a hint, so there are some cases where the service cannot adhere to these conditions strictly. For example, record boundaries might be such that the size is a little over or under the configured buffering size. By default, no encryption is performed. We strongly recommend that you enable encryption to ensure secure data storage in Amazon S3.

A few notes about Amazon Redshift as a destination:

An Amazon Redshift destination requires an S3 bucket as intermediate location. Kinesis Data Firehose first delivers data to Amazon S3 and then uses COPY syntax to load data into an Amazon Redshift table. This is specified in the RedshiftDestinationConfiguration.S3Configuration parameter.

The compression formats SNAPPY or ZIP cannot be specified in RedshiftDestinationConfiguration.S3Configuration because the Amazon Redshift COPY operation that reads from the S3 bucket doesn't support these compression formats.

We strongly recommend that you use the user name and password you provide exclusively with Kinesis Data Firehose, and that the permissions for the account are restricted for Amazon Redshift INSERT permissions.

Kinesis Data Firehose assumes the IAM role that is configured as part of the destination. The role should allow the Kinesis Data Firehose principal to assume the role, and the role should have permissions that allow the service to deliver the data. For more information, see Grant Kinesis Data Firehose Access to an Amazon S3 Destination in the Amazon Kinesis Data Firehose Developer Guide.

Implementations

impl<C, M, R> CreateDeliveryStream<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateDeliveryStreamOutput, SdkError<CreateDeliveryStreamError>> where
R::Policy: SmithyRetryPolicy<CreateDeliveryStreamInputOperationOutputAlias, CreateDeliveryStreamOutput, CreateDeliveryStreamError, CreateDeliveryStreamInputOperationRetryAlias>,

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn delivery_stream_name(self, inp: impl Into<String>) -> Self

The name of the delivery stream. This name must be unique per AWS account in the same AWS Region. If the delivery streams are in different accounts or different Regions, you can have multiple delivery streams with the same name.

pub fn set_delivery_stream_name(self, input: Option<String>) -> Self

The name of the delivery stream. This name must be unique per AWS account in the same AWS Region. If the delivery streams are in different accounts or different Regions, you can have multiple delivery streams with the same name.

pub fn delivery_stream_type(self, inp: DeliveryStreamType) -> Self

The delivery stream type. This parameter can be one of the following values:

DirectPut: Provider applications access the delivery stream directly.

KinesisStreamAsSource: The delivery stream uses a Kinesis data stream as a source.

pub fn set_delivery_stream_type(self, input: Option<DeliveryStreamType>) -> Self

The delivery stream type. This parameter can be one of the following values:

DirectPut: Provider applications access the delivery stream directly.

KinesisStreamAsSource: The delivery stream uses a Kinesis data stream as a source.

pub fn kinesis_stream_source_configuration(
self,
inp: KinesisStreamSourceConfiguration
) -> Self

When a Kinesis data stream is used as the source for the delivery stream, a KinesisStreamSourceConfiguration containing the Kinesis data stream Amazon Resource Name (ARN) and the role ARN for the source stream.

pub fn set_kinesis_stream_source_configuration(
self,
input: Option<KinesisStreamSourceConfiguration>
) -> Self

When a Kinesis data stream is used as the source for the delivery stream, a KinesisStreamSourceConfiguration containing the Kinesis data stream Amazon Resource Name (ARN) and the role ARN for the source stream.

pub fn delivery_stream_encryption_configuration_input(
self,
inp: DeliveryStreamEncryptionConfigurationInput
) -> Self

Used to specify the type and Amazon Resource Name (ARN) of the KMS key needed for Server-Side Encryption (SSE).

pub fn set_delivery_stream_encryption_configuration_input(
self,
input: Option<DeliveryStreamEncryptionConfigurationInput>
) -> Self

Used to specify the type and Amazon Resource Name (ARN) of the KMS key needed for Server-Side Encryption (SSE).

pub fn s3_destination_configuration(
self,
inp: S3DestinationConfiguration
) -> Self

[Deprecated] The destination in Amazon S3. You can specify only one destination.

pub fn set_s3_destination_configuration(
self,
input: Option<S3DestinationConfiguration>
) -> Self

[Deprecated] The destination in Amazon S3. You can specify only one destination.

pub fn extended_s3_destination_configuration(
self,
inp: ExtendedS3DestinationConfiguration
) -> Self

The destination in Amazon S3. You can specify only one destination.

pub fn set_extended_s3_destination_configuration(
self,
input: Option<ExtendedS3DestinationConfiguration>
) -> Self

The destination in Amazon S3. You can specify only one destination.

pub fn redshift_destination_configuration(
self,
inp: RedshiftDestinationConfiguration
) -> Self

The destination in Amazon Redshift. You can specify only one destination.

pub fn set_redshift_destination_configuration(
self,
input: Option<RedshiftDestinationConfiguration>
) -> Self

The destination in Amazon Redshift. You can specify only one destination.

pub fn elasticsearch_destination_configuration(
self,
inp: ElasticsearchDestinationConfiguration
) -> Self

The destination in Amazon ES. You can specify only one destination.

pub fn set_elasticsearch_destination_configuration(
self,
input: Option<ElasticsearchDestinationConfiguration>
) -> Self

The destination in Amazon ES. You can specify only one destination.

pub fn amazonopensearchservice_destination_configuration(
self,
inp: AmazonopensearchserviceDestinationConfiguration
) -> Self

pub fn set_amazonopensearchservice_destination_configuration(
self,
input: Option<AmazonopensearchserviceDestinationConfiguration>
) -> Self

pub fn splunk_destination_configuration(
self,
inp: SplunkDestinationConfiguration
) -> Self

The destination in Splunk. You can specify only one destination.

pub fn set_splunk_destination_configuration(
self,
input: Option<SplunkDestinationConfiguration>
) -> Self

The destination in Splunk. You can specify only one destination.

pub fn http_endpoint_destination_configuration(
self,
inp: HttpEndpointDestinationConfiguration
) -> Self

Enables configuring Kinesis Firehose to deliver data to any HTTP endpoint destination. You can specify only one destination.

pub fn set_http_endpoint_destination_configuration(
self,
input: Option<HttpEndpointDestinationConfiguration>
) -> Self

Enables configuring Kinesis Firehose to deliver data to any HTTP endpoint destination. You can specify only one destination.

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

A set of tags to assign to the delivery stream. A tag is a key-value pair that you can define and assign to AWS resources. Tags are metadata. For example, you can add friendly names and descriptions or other types of information that can help you distinguish the delivery stream. For more information about tags, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

You can specify up to 50 tags when creating a delivery stream.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

A set of tags to assign to the delivery stream. A tag is a key-value pair that you can define and assign to AWS resources. Tags are metadata. For example, you can add friendly names and descriptions or other types of information that can help you distinguish the delivery stream. For more information about tags, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide.

You can specify up to 50 tags when creating a delivery stream.