Module aws_sdk_firehose::error

Modules

concurrent_modification_exception

See ConcurrentModificationException

invalid_argument_exception

See InvalidArgumentException

invalid_kms_resource_exception

See InvalidKmsResourceException

limit_exceeded_exception

See LimitExceededException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

Structs

ConcurrentModificationException

Another modification has already happened. Fetch VersionId again and use

CreateDeliveryStreamError

DeleteDeliveryStreamError

DescribeDeliveryStreamError

InvalidArgumentException

The specified input parameter has a value that is not valid.

InvalidKmsResourceException

Kinesis Data Firehose throws this exception when an attempt to put records or to start

LimitExceededException

You have already reached the limit for a requested resource.

ListDeliveryStreamsError

ListTagsForDeliveryStreamError

PutRecordBatchError

PutRecordError

ResourceInUseException

The resource is already in use and not available for this operation.

ResourceNotFoundException

The specified resource could not be found.

ServiceUnavailableException

The service is unavailable. Back off and retry the operation. If you continue to see

StartDeliveryStreamEncryptionError

StopDeliveryStreamEncryptionError

TagDeliveryStreamError

UntagDeliveryStreamError

UpdateDestinationError

Enums

CreateDeliveryStreamErrorKind

DeleteDeliveryStreamErrorKind

DescribeDeliveryStreamErrorKind

ListDeliveryStreamsErrorKind

ListTagsForDeliveryStreamErrorKind

PutRecordBatchErrorKind

PutRecordErrorKind

StartDeliveryStreamEncryptionErrorKind

StopDeliveryStreamEncryptionErrorKind

TagDeliveryStreamErrorKind

UntagDeliveryStreamErrorKind

UpdateDestinationErrorKind