Module aws_sdk_firehose::model

Modules

buffering_hints

See BufferingHints

cloud_watch_logging_options

See CloudWatchLoggingOptions

copy_command

See CopyCommand

data_format_conversion_configuration

See DataFormatConversionConfiguration

delivery_stream_description

See DeliveryStreamDescription

delivery_stream_encryption_configuration

See DeliveryStreamEncryptionConfiguration

delivery_stream_encryption_configuration_input

See DeliveryStreamEncryptionConfigurationInput

deserializer

See Deserializer

destination_description

See DestinationDescription

dynamic_partitioning_configuration

See DynamicPartitioningConfiguration

elasticsearch_buffering_hints

See ElasticsearchBufferingHints

elasticsearch_destination_configuration

See ElasticsearchDestinationConfiguration

elasticsearch_destination_description

See ElasticsearchDestinationDescription

elasticsearch_destination_update

See ElasticsearchDestinationUpdate

elasticsearch_retry_options

See ElasticsearchRetryOptions

encryption_configuration

See EncryptionConfiguration

extended_s3_destination_configuration

See ExtendedS3DestinationConfiguration

extended_s3_destination_description

See ExtendedS3DestinationDescription

extended_s3_destination_update

See ExtendedS3DestinationUpdate

failure_description

See FailureDescription

hive_json_ser_de

See HiveJsonSerDe

http_endpoint_buffering_hints

See HttpEndpointBufferingHints

http_endpoint_common_attribute

See HttpEndpointCommonAttribute

http_endpoint_configuration

See HttpEndpointConfiguration

http_endpoint_description

See HttpEndpointDescription

http_endpoint_destination_configuration

See HttpEndpointDestinationConfiguration

http_endpoint_destination_description

See HttpEndpointDestinationDescription

http_endpoint_destination_update

See HttpEndpointDestinationUpdate

http_endpoint_request_configuration

See HttpEndpointRequestConfiguration

http_endpoint_retry_options

See HttpEndpointRetryOptions

input_format_configuration

See InputFormatConfiguration

kinesis_stream_source_configuration

See KinesisStreamSourceConfiguration

kinesis_stream_source_description

See KinesisStreamSourceDescription

kms_encryption_config

See KmsEncryptionConfig

open_x_json_ser_de

See OpenXJsonSerDe

orc_ser_de

See OrcSerDe

output_format_configuration

See OutputFormatConfiguration

parquet_ser_de

See ParquetSerDe

processing_configuration

See ProcessingConfiguration

processor

See Processor

processor_parameter

See ProcessorParameter

put_record_batch_response_entry

See PutRecordBatchResponseEntry

record

See Record

redshift_destination_configuration

See RedshiftDestinationConfiguration

redshift_destination_description

See RedshiftDestinationDescription

redshift_destination_update

See RedshiftDestinationUpdate

redshift_retry_options

See RedshiftRetryOptions

retry_options

See RetryOptions

s3_destination_configuration

See S3DestinationConfiguration

s3_destination_description

See S3DestinationDescription

s3_destination_update

See S3DestinationUpdate

schema_configuration

See SchemaConfiguration

serializer

See Serializer

source_description

See SourceDescription

splunk_destination_configuration

See SplunkDestinationConfiguration

splunk_destination_description

See SplunkDestinationDescription

splunk_destination_update

See SplunkDestinationUpdate

splunk_retry_options

See SplunkRetryOptions

tag

See Tag

vpc_configuration

See VpcConfiguration

vpc_configuration_description

See VpcConfigurationDescription

Structs

BufferingHints

Describes hints for the buffering to perform before delivering data to the

CloudWatchLoggingOptions

Describes the Amazon CloudWatch logging options for your delivery stream.

CopyCommand

Describes a COPY command for Amazon Redshift.

DataFormatConversionConfiguration

Specifies that you want Kinesis Data Firehose to convert data from the JSON format to

DeliveryStreamDescription

Contains information about a delivery stream.

DeliveryStreamEncryptionConfiguration

Contains information about the server-side encryption (SSE) status for the delivery

DeliveryStreamEncryptionConfigurationInput

Specifies the type and Amazon Resource Name (ARN) of the CMK to use for Server-Side

Deserializer

The deserializer you want Kinesis Data Firehose to use for converting the input data

DestinationDescription

Describes the destination for a delivery stream.

DynamicPartitioningConfiguration

The configuration of the dynamic partitioning mechanism that creates smaller data sets

ElasticsearchBufferingHints

Describes the buffering to perform before delivering data to the Amazon ES

ElasticsearchDestinationConfiguration

Describes the configuration of a destination in Amazon ES.

ElasticsearchDestinationDescription

The destination description in Amazon ES.

ElasticsearchDestinationUpdate

Describes an update for a destination in Amazon ES.

ElasticsearchRetryOptions

Configures retry behavior in case Kinesis Data Firehose is unable to deliver

EncryptionConfiguration

Describes the encryption for a destination in Amazon S3.

ExtendedS3DestinationConfiguration

Describes the configuration of a destination in Amazon S3.

ExtendedS3DestinationDescription

Describes a destination in Amazon S3.

ExtendedS3DestinationUpdate

Describes an update for a destination in Amazon S3.

FailureDescription

Provides details in case one of the following operations fails due to an error related

HiveJsonSerDe

The native Hive / HCatalog JsonSerDe. Used by Kinesis Data Firehose for deserializing

HttpEndpointBufferingHints

Describes the buffering options that can be applied before data is delivered to the HTTP

HttpEndpointCommonAttribute

Describes the metadata that's delivered to the specified HTTP endpoint

HttpEndpointConfiguration

Describes the configuration of the HTTP endpoint to which Kinesis Firehose delivers

HttpEndpointDescription

Describes the HTTP endpoint selected as the destination.

HttpEndpointDestinationConfiguration

Describes the configuration of the HTTP endpoint destination.

HttpEndpointDestinationDescription

Describes the HTTP endpoint destination.

HttpEndpointDestinationUpdate

Updates the specified HTTP endpoint destination.

HttpEndpointRequestConfiguration

The configuration of the HTTP endpoint request.

HttpEndpointRetryOptions

Describes the retry behavior in case Kinesis Data Firehose is unable to deliver data to

InputFormatConfiguration

Specifies the deserializer you want to use to convert the format of the input data.

KinesisStreamSourceConfiguration

The stream and role Amazon Resource Names (ARNs) for a Kinesis data stream used as

KinesisStreamSourceDescription

Details about a Kinesis data stream used as the source for a Kinesis Data Firehose

KmsEncryptionConfig

Describes an encryption key for a destination in Amazon S3.

OpenXJsonSerDe

The OpenX SerDe. Used by Kinesis Data Firehose for deserializing data, which means

OrcSerDe

A serializer to use for converting data to the ORC format before storing it in Amazon

OutputFormatConfiguration

Specifies the serializer that you want Kinesis Data Firehose to use to convert the

ParquetSerDe

A serializer to use for converting data to the Parquet format before storing it in

ProcessingConfiguration

Describes a data processing configuration.

Processor

Describes a data processor.

ProcessorParameter

Describes the processor parameter.

PutRecordBatchResponseEntry

Contains the result for an individual record from a PutRecordBatch

Record

The unit of data in a delivery stream.

RedshiftDestinationConfiguration

Describes the configuration of a destination in Amazon Redshift.

RedshiftDestinationDescription

Describes a destination in Amazon Redshift.

RedshiftDestinationUpdate

Describes an update for a destination in Amazon Redshift.

RedshiftRetryOptions

Configures retry behavior in case Kinesis Data Firehose is unable to deliver

RetryOptions

The retry behavior in case Kinesis Data Firehose is unable to deliver data to an Amazon

S3DestinationConfiguration

Describes the configuration of a destination in Amazon S3.

S3DestinationDescription

Describes a destination in Amazon S3.

S3DestinationUpdate

Describes an update for a destination in Amazon S3.

SchemaConfiguration

Specifies the schema to which you want Kinesis Data Firehose to configure your data

Serializer

The serializer that you want Kinesis Data Firehose to use to convert data to the target

SourceDescription

Details about a Kinesis data stream used as the source for a Kinesis Data Firehose

SplunkDestinationConfiguration

Describes the configuration of a destination in Splunk.

SplunkDestinationDescription

Describes a destination in Splunk.

SplunkDestinationUpdate

Describes an update for a destination in Splunk.

SplunkRetryOptions

Configures retry behavior in case Kinesis Data Firehose is unable to deliver

Tag

Metadata that you can assign to a delivery stream, consisting of a key-value

VpcConfiguration

The details of the VPC of the Amazon ES destination.

VpcConfigurationDescription

The details of the VPC of the Amazon ES destination.

Enums

CompressionFormat

ContentEncoding

DeliveryStreamEncryptionStatus

DeliveryStreamFailureType

DeliveryStreamStatus

DeliveryStreamType

ElasticsearchIndexRotationPeriod

ElasticsearchS3BackupMode

HecEndpointType

HttpEndpointS3BackupMode

KeyType

NoEncryptionConfig

OrcCompression

OrcFormatVersion

ParquetCompression

ParquetWriterVersion

ProcessorParameterName

ProcessorType

RedshiftS3BackupMode

S3BackupMode

SplunkS3BackupMode