Struct aws_sdk_firehose::model::s3_destination_configuration::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for S3DestinationConfiguration

Implementations

impl Builder

pub fn role_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the AWS credentials. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

pub fn set_role_arn(self, input: Option<String>) -> Self

pub fn bucket_arn(self, input: impl Into<String>) -> Self

The ARN of the S3 bucket. For more information, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

pub fn set_bucket_arn(self, input: Option<String>) -> Self

pub fn prefix(self, input: impl Into<String>) -> Self

The "YYYY/MM/DD/HH" time format prefix is automatically used for delivered Amazon S3 files. You can also specify a custom prefix, as described in Custom Prefixes for Amazon S3 Objects.

pub fn set_prefix(self, input: Option<String>) -> Self

pub fn error_output_prefix(self, input: impl Into<String>) -> Self

A prefix that Kinesis Data Firehose evaluates and adds to failed records before writing them to S3. This prefix appears immediately following the bucket name. For information about how to specify this prefix, see Custom Prefixes for Amazon S3 Objects.

pub fn set_error_output_prefix(self, input: Option<String>) -> Self

pub fn buffering_hints(self, input: BufferingHints) -> Self

The buffering option. If no value is specified, BufferingHints object default values are used.

pub fn set_buffering_hints(self, input: Option<BufferingHints>) -> Self

pub fn compression_format(self, input: CompressionFormat) -> Self

The compression format. If no value is specified, the default is UNCOMPRESSED.

The compression formats SNAPPY or ZIP cannot be specified for Amazon Redshift destinations because they are not supported by the Amazon Redshift COPY operation that reads from the S3 bucket.

pub fn set_compression_format(self, input: Option<CompressionFormat>) -> Self

pub fn encryption_configuration(self, input: EncryptionConfiguration) -> Self

The encryption configuration. If no value is specified, the default is no encryption.

pub fn set_encryption_configuration(
self,
input: Option<EncryptionConfiguration>
) -> Self

pub fn cloud_watch_logging_options(
self,
input: CloudWatchLoggingOptions
) -> Self

The CloudWatch logging options for your delivery stream.

pub fn set_cloud_watch_logging_options(
self,
input: Option<CloudWatchLoggingOptions>
) -> Self

pub fn build(self) -> S3DestinationConfiguration

Consumes the builder and constructs a S3DestinationConfiguration